requirements.txt
argh
paramiko
PyYAML
